BankStatus ADT, opaque types, named results in Banking.scala (#94)
* Refactor Banking.scala to Firm/Household quality level

BankStatus ADT (Active/Failed) replaces 3 flat Boolean/Int fields.
Named constants for all Basel III weights and calibration numbers.
Opaque types: nplRatio/car/lcr/nsfr → Ratio, hqla/asf/rsf → PLN,
method params PLN/Rate where appropriate.
Named result types: PerBankAmounts, FailureCheckResult, BailInResult,
ResolutionResult replace raw tuples.
BankState fields have no defaults — all constructors explicit.
Inline field comments matching Firm.State/Household.State pattern.
Test mkBank helpers reduce constructor boilerplate.

31 files changed, 1246 tests pass.

* Eliminate all vars from Banking.scala

assignBank: while loop → scanLeft + indexWhere
checkFailures: var anyFailed → prevFailed set + exists
applyBailIn: var totalBailIn → map to (bank, haircut) pairs + kahanSum
allocateBonds/allocateQePurchases: var allocated → foldLeft with accumulator

* Remove default values from Banking.initialize parameters

All 6 params now explicit — no hidden zeros. Aligns with
BankState (no defaults) convention.

* Remove default from State.interbankCurve

* Remove defaults, dead param, deduplicate filter in Banking

- canLend/checkFailures: remove default ccyb — all callers explicit
- clearInterbank: remove unused rate parameter
- resolveFailures: single-pass foldLeft replaces 6 × kahanSumBy
- interbankRate: single filterNot(_.failed) instead of two

+++ src/main/scala/sfc/SimOutput.scala
@@ -124,7 +124,7 @@ object SimOutput:
     ColumnDef("ExRate", ctx => ctx.world.forex.exchangeRate),
     ColumnDef("MarketWage", ctx => ctx.world.hhAgg.marketWage.toDouble),
     ColumnDef("GovDebt", ctx => ctx.world.gov.cumulativeDebt.toDouble),
-    ColumnDef("NPL", ctx => ctx.world.bankingSector.aggregate.nplRatio),
+    ColumnDef("NPL", ctx => ctx.world.bankingSector.aggregate.nplRatio.toDouble),
     ColumnDef("RefRate", ctx => ctx.world.nbp.referenceRate.toDouble),
     ColumnDef("PriceLevel", ctx => ctx.world.priceLevel),
     ColumnDef("AutoRatio", ctx => ctx.world.real.automationRatio.toDouble),
@@ -174,8 +174,8 @@ object SimOutput:
     ColumnDef("FxInterventionActive", ctx => if ctx.p.flags.nbpFxIntervention then 1.0 else 0.0),
     // Interbank
     ColumnDef("InterbankRate", ctx => ctx.world.bankingSector.interbankRate.toDouble),
-    ColumnDef("MinBankCAR", ctx => if ctx.aliveBanks.isEmpty then 0.0 else ctx.aliveBanks.map(_.car).min),
-    ColumnDef("MaxBankNPL", ctx => if ctx.aliveBanks.isEmpty then 0.0 else ctx.aliveBanks.map(_.nplRatio).max),
+    ColumnDef("MinBankCAR", ctx => if ctx.aliveBanks.isEmpty then 0.0 else ctx.aliveBanks.map(_.car).min.toDouble),
+    ColumnDef("MaxBankNPL", ctx => if ctx.aliveBanks.isEmpty then 0.0 else ctx.aliveBanks.map(_.nplRatio).max.toDouble),
     ColumnDef("BankFailures", ctx => ctx.world.bankingSector.banks.count(_.failed).toDouble),
     // Monetary plumbing — now read from World
     ColumnDef("ReserveInterest", ctx => ctx.world.plumbing.reserveInterestTotal.toDouble),
@@ -195,9 +195,9 @@ object SimOutput:
     // LCR/NSFR
     ColumnDef(
       "MinBankLCR",
-      ctx => { given SimParams = ctx.p; if ctx.aliveBanks.isEmpty then 0.0 else ctx.aliveBanks.map(_.lcr).min },
+      ctx => { given SimParams = ctx.p; if ctx.aliveBanks.isEmpty then 0.0 else ctx.aliveBanks.map(_.lcr).min.toDouble },
     ),
-    ColumnDef("MinBankNSFR", ctx => if ctx.aliveBanks.isEmpty then 0.0 else ctx.aliveBanks.map(_.nsfr).min),
+    ColumnDef("MinBankNSFR", ctx => if ctx.aliveBanks.isEmpty then 0.0 else ctx.aliveBanks.map(_.nsfr).min.toDouble),
     ColumnDef(
       "AvgTermDepositFrac",
       ctx =>
